This update has both the fixityservice and bitintegritytoolsservice place service output files in the system space, 'x-service-out'. It also has those services create output files with a naming convention that is predictable yet avoids clashes. The names have the 'bitintegrity' prefix and contain a combination of the target spaceId and the date, such as: x-service-out : bitintegrity/fingerprints-awoods-images-2010-12-07.csv
Additionally, a new project was created for fixityservicetools, named bitintegritytoolsservice.
The 'tools' profile in fixityservice was being used to create a service package that exposes just the extra tools from the fixityservice project. Everything works well except in the OSGi configuration (serviceutil.DuraConfigAdminImpl). When this 'tools' service is deployed before the standard fixityservice, the fixityservice is unable to get a proper handle to its configuration map. The ConfigAdmin assigns ownership of the fixityservice config map to the servicesutil bundle. The workaround is the creation of the separate bitintegritytoolsservice project.



git-svn-id: https://svn.duraspace.org/duracloud/trunk@285 1005ed41-97cd-4a8f-848c-be5b5fe45bcb