This update resolves https://jira.duraspace.org/browse/DURACLOUD-286 . It both hides system services (webapputil & imagemagick) as well as deploys them when deploying j2kservice and imageconversion, respectively. Likewise, the dependency services are undeployed when their parent services are undeployed. In order to achieve this, the services.xml schema was updated to hold two new values per service-type:  boolean:isSystemService  list:serviceDependencies
The new version of the schema is 1.0.0, up from 0.8.0.

After some initial testing it became clear that wrapping the dependency service into the parent service packages was not feasible.
Additionally, app-config was updated to not initialize durareport if the durareport configuration does not exist in the init.properties.
The full store/space/content metadata collection durareport performs was using up all of the machine's file resources.



git-svn-id: https://svn.duraspace.org/duracloud/trunk@416 1005ed41-97cd-4a8f-848c-be5b5fe45bcb