Applies Danny's patch to the bug related to https://jira.duraspace.org/browse/DURACLOUD-710.
From Danny:  The listener on the recount button was being added each time a space was clicked on.  This was causing  ever increasing numbers of ajax calls to queue up every time a space panel was loaded.  When clicking on  the recount then,  the different ajax calls were returning not necessarily in the order they were served on the server side thus accounting for both noticeable server load and counts bouncing all over the place. The fix simply clears any attached listeners before adding a new one.

git-svn-id: https://svn.duraspace.org/duracloud/trunk@981 1005ed41-97cd-4a8f-848c-be5b5fe45bcb