Completes: https://jira.duraspace.org/browse/DURACLOUD-774
Combines the SyncTool and UpSync into a single executable jar. This jar is a build artifact of the synctoolui project, named duracloud-synctool-<version>.jar. Running this jar with command line arguments works just like the SyncTool. Running the jar with no arguments launches UpSync.

Part of this work involved transitioning the SyncTool to use logback for logging rather than log4j (and thus removing the feature which allowed the user to change the log level on the command line at run time). The work directory parameter of the Sync Tool is also now optional. If it is not provided, a default value (in the user's home directory) is used. This default value is also the value used when UpSync is put in use.

git-svn-id: https://svn.duraspace.org/duracloud/trunk@1072 1005ed41-97cd-4a8f-848c-be5b5fe45bcb