This commit ensures transfer of chunk manifest is retried on failure.
Additionally, I added missing retry behavior when the DuracloudContentWriter
is in no throw mode.  Also, it seemed to me suboptimal to have the DuracloudContentWriter
continue uploading chunks even after multiple retry failures while in
no throw mode. Chunks and manifests were being uploaded even if
individual chunks failed.  Now that retries are in place and now that
we are checking for existing chunks before transferring them,  I think
it is cleaner not to upload the manifest until the set of chunks
has been completely uploaded when the DuracloudContentWriter is in
"no throw" mode.  When retries were not happening in the no throw mode,
I can see how a user might want to try to get as much of the file transferred
as possible and then deal with the failed chunks separately.  However,
until very recently,  existing chunks were not being checked, so in the
case of failure of a single chunk, the entire file was being re-transferred.
Now that retries have been added, it is much less likely that a content
writer will get tripped and fail.   If a single chunk fails after multiple
retries, it seems better that the write operation should fail quickly.

Resolves: https://jira.duraspace.org/browse/DURACLOUD-1082