Add S3 Swift storage provider
This commit adds a new storage option - OpenStack Swift with S3
middleware.

We have written a new module called swiftstorageprovider which overrides
the methods in s3storageprovider that Swift does not support.

S3ProviderUtil is now capable of constructing S3 clients compatible with
both AWS and Swift, and handles the options passed in for Swift clients.