DURACLOUD-1227: Adds a new S3 Swift storage provider (PR #96)
This commit adds a new storage option - OpenStack Swift with S3 middleware.

We have written a new module called swiftstorageprovider which overrides the methods in s3storageprovider that Swift does not support. Space properties are stored as text objects within a hidden space.

S3ProviderUtil is now capable of constructing S3 clients compatible with both AWS and Swift, and handles the options passed in for Swift clients.