DURACLOUD-1313: Stops synctool retrying hundreds of times on missing space (#154)
* Bump hibernate version

In order to overcome this issue.
https://github.com/spring-projects/spring-boot/issues/11225

Also had to pull back jaxb-runtime, as this is now pinned by hibernate.

* Stop syncclient from retrying hundreds of times

We were previously creating 3 nested retry loops, first in
DuraStoreSyncEndpoint, then again in ContentIterator, ultimately
calling the `getSpace` method, which is already a `Retriable`! We
now solely rely on the retry loop created by `ContentStore.getSpace()`.

* Revert ContentIterator and change spaceExists method

Reverted ContentIterator.

Switched out the call in spaceExists() to use getSpaceACLs() instead of
getSpaceContents() so it doesn't end up in a double loop when the space
does not exist (and therefore returns an error faster).

* Don't capture acls in a variable

There is no need to do this because the return value is never used.

* Do not create space if it does not exist

When running in command-line mode, it is too easy to mis-type the space
name. Next thing you know, you've got tons of content in the incorrect
space. Synctool will now error if the space does not exist.

Co-authored-by: Andy Foster <andy.foster@utoronto.ca>